server/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
streamlit
google-genai

